requirements.unlocked.txt
python-slugify
pyyaml
scrapy>=1.1.0rc
